feat: implement cpu utilization

+++ src/processor.cpp
@@ -5,5 +5,5 @@
 #include<string>
 
 float Processor::Utilization() {
-   return LinuxParser::ActiveJiffies(this->cpu_stats_file_path_) / LinuxParser::Jiffies(this->cpu_stats_file_path_);
+   return (float) LinuxParser::ActiveJiffies(this->cpu_stats_file_path_) / (float) LinuxParser::Jiffies(this->cpu_stats_file_path_);
 }
